pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rsgen8k"
version = "0.1.0"
description = "Remote Sensing Image Generation Scaling to 8K Resolution via MegaFusion"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.28.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "datasets>=2.14.0",
    "einops>=0.7.0",
    "Pillow>=9.0.0",
    "pyyaml>=6.0",
    "safetensors>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
xformers = [
    "xformers>=0.0.20",
]

[project.scripts]
rsgen8k-generate = "rsgen8k.generate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
